requirements.txt
fastapi
uvicorn
requests
aiohttp
beautifulsoup4
sentence-transformers
torch
faiss-cpu
chromadb
pandas
numpy
nltk
spacy
openai
httpx
python-dotenv
pytest
pytest-asyncio
